+++ CYD-Klipper/src/core/printer_integration.hpp
@@ -159,7 +159,7 @@ class BasePrinter
         virtual Files get_files() = 0;
         virtual bool start_file(const char* filename) = 0;
         virtual unsigned char* get_32_32_png_image_thumbnail(const char* gcode_filename);
-        virtual bool set_target_temperature(PrinterTemperatureDevice device, float temperature) = 0;
+        virtual bool set_target_temperature(PrinterTemperatureDevice device, unsigned int temperature) = 0;
 
         BasePrinter(unsigned char index);
         PrinterData* AnnouncePrinterData();
